requirements.txt
flask==1.1.2
pymysql==1.0.2
python-dotenv==0.15.0